Add files via upload
This build has a completely working interface and I changed the color scheme such that the maze and the player are a different color. Now the maze is red, the player is yellow, and the buttons are black. I was considering making the buttons and the player the same color so that the implication is that the  buttons control the player, but I would rather have it that the button is black